+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Created Campus Guide home widget [#1808](https://github.com/rokwire/illinois-app/issues/1808).
+
 ## [4.0.35] - 2022-06-22
 ### Fixed
 - Fixed section toggle from Browse panel [#1798](https://github.com/rokwire/illinois-app/issues/1798).